removing long double from test_concurrent_1d test as it fails with new gnuplot-iostream

+++ tests/concurrent_1d/test_concurrent_1d.cpp
@@ -30,7 +30,7 @@ int main()
 //<listing-1>
   struct ct_params_t : ct_params_default_t
   { 
-    using real_t = long double; 
+    using real_t = double; 
     enum { n_dims = 1 }; 
     enum { n_eqns = 1 };
   };
